Added ImageImport type
Image imports happen now by means of new ImageImport objects and not by
bumping the generation value in a Tag (that is now known as Image).
This commit adds a type called ImageImport that will be instantiated
whenever a new image import is necessary.

This PR also decreases the scope of this project:

1. No more quay webhook handler
2. No more docker hub webhook handler
3. No more deployment integration

Kubectl plugin is now renamed to kubectl-image (instead of kubectl-tag).